+++ minijinja/src/value.rs
@@ -1523,6 +1523,34 @@ impl<'a> fmt::Debug for ValueIterator {
     }
 }
 
+/// Helper to pass a single value as context bound to a name.
+///
+/// The first argument is the name of the variable, the second is the
+/// value to pass.  Here a simple motivating example:
+///
+/// ```
+/// use minijinja::{Environment, value::Single};
+///
+/// let mut env = Environment::new();
+/// env.add_template("simple", "Hello {{ name }}!").unwrap();
+/// let tmpl = env.get_template("simple").unwrap();
+/// let rv = tmpl.render(Single("name", "Peter")).unwrap();
+/// ```
+#[derive(Debug, Clone)]
+pub struct Single<'a, V: Serialize + ?Sized>(pub &'a str, pub V);
+
+impl<'a, V: Serialize + ?Sized> Serialize for Single<'a, V> {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        use serde::ser::SerializeMap;
+        let mut map = serializer.serialize_map(Some(1))?;
+        map.serialize_entry(self.0, &self.1)?;
+        map.end()
+    }
+}
+
 enum ValueIteratorImpl<'a> {
     Empty,
     Seq(std::slice::Iter<'a, Value>),